chore: Extract YamlParser
- Some cleanup + renaming
