+++ config/detekt/config.yml
@@ -7,7 +7,7 @@ build:
 
 style:
   MagicNumber:
-    # rule is to sensitive (e.g. you can't define a for loop without using const for from, to and stepSize)
+    # Rule is to sensitive (e.g. you can't define a for loop without using const for from, to and stepSize)
     ignoreRanges: true
     ignoreEnums: true
     ignorePropertyDeclaration: true # Compose
@@ -18,10 +18,12 @@ style:
     active: false
   ReturnCount:
     excludeGuardClauses: true
+    # It is fine to have TO-DO statements
   ForbiddenComment:
     active: false
 
 exceptions:
+  # Sometimes you want to bundle multiple exceptions to be caught the same way
   TooGenericExceptionCaught:
     active: false
